Add problem category to medicine

+++ featureMedicine/src/main/java/cz/vvoleman/phr/featureMedicine/ui/component/timeSelector/TimeAdapter.kt
@@ -63,7 +63,7 @@ class TimeAdapter(
                                 val newValue = if (it.isNullOrBlank()) {
                                     0
                                 } else {
-                                    it.toString().toInt()
+                                    it.toString().toFloat()
                                 }
 
                                 listener.onQuantityChange(
